Move emitting to meta_metrics
