+++ pkg/txm/clientwrappers/dualbroadcast/meta_client.go
@@ -17,17 +17,13 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	evmtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/gogo/protobuf/proto"
-	"google.golang.org/protobuf/encoding/protojson"
 
 	"github.com/mitchellh/mapstructure"
 
-	"github.com/smartcontractkit/chainlink-common/pkg/beholder"
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 
 	"github.com/smartcontractkit/chainlink-evm/pkg/txm"
 	"github.com/smartcontractkit/chainlink-evm/pkg/txm/types"
-	atlaspb "github.com/smartcontractkit/chainlink-protos/svr/v1"
 )
 
 const (
@@ -135,31 +131,27 @@ type MetaClientRPC interface {
 }
 
 type MetaClient struct {
-	lggr            logger.SugaredLogger
-	c               MetaClientRPC
-	ks              MetaClientKeystore
-	customURL       *url.URL
-	chainID         *big.Int
-	metrics         *MetaMetrics
-	beholderEmitter beholder.Emitter
+	lggr      logger.SugaredLogger
+	c         MetaClientRPC
+	ks        MetaClientKeystore
+	customURL *url.URL
+	chainID   *big.Int
+	metrics   *MetaMetrics
 }
 
-const schemaBasePath = "/oev-fastlane-atlas-error/versions/1"
-
 func NewMetaClient(lggr logger.Logger, c MetaClientRPC, ks MetaClientKeystore, customURL *url.URL, chainID *big.Int) (*MetaClient, error) {
-	metrics, err := NewMetaMetrics(chainID.String())
+	metrics, err := NewMetaMetrics(chainID.String(), lggr)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create Meta metrics: %w", err)
 	}
 
 	return &MetaClient{
-		lggr:            logger.Sugared(logger.Named(lggr, "Txm.MetaClient")),
-		c:               c,
-		ks:              ks,
-		customURL:       customURL,
-		chainID:         chainID,
-		metrics:         metrics,
-		beholderEmitter: beholder.GetEmitter(),
+		lggr:      logger.Sugared(logger.Named(lggr, "Txm.MetaClient")),
+		c:         c,
+		ks:        ks,
+		customURL: customURL,
+		chainID:   chainID,
+		metrics:   metrics,
 	}, nil
 }
 
@@ -169,69 +161,6 @@ func (a *MetaClient) NonceAt(ctx context.Context, address common.Address, blockN
 
 func (a *MetaClient) PendingNonceAt(ctx context.Context, address common.Address) (uint64, error) {
 	return a.c.PendingNonceAt(ctx, address)
-}
-
-// TODO(gg): move this to meta_metrics.go
-// emitAtlasErrorWithHttpStatusCode emits an OTel event to track FastLane Atlas errors with an HTTP status code
-func (a *MetaClient) emitAtlasErrorWithHttpStatusCode(ctx context.Context, errType string, cause error, httpStatusCode int, tx *types.Transaction) {
-	var nonce string
-	if tx.Nonce != nil {
-		nonce = fmt.Sprintf("%d", *tx.Nonce)
-	}
-
-	meta, err := tx.GetMeta()
-	if err != nil {
-		a.lggr.Errorw(fmt.Sprintf("Failed to get meta for tx. Error to emit was: %v", cause), "txId", tx.ID, "err", err)
-		return
-	}
-
-	var destAddress string
-	if meta != nil && meta.FwdrDestAddress != nil {
-		destAddress = meta.FwdrDestAddress.String()
-	}
-
-	msg := &atlaspb.FastLaneAtlasError{
-		ChainId:        a.chainID.String(),
-		FromAddress:    tx.FromAddress.Hex(),
-		ToAddress:      tx.ToAddress.Hex(),
-		FeedAddress:    destAddress,
-		Nonce:          nonce,
-		ErrorType:      errType,
-		ErrorMessage:   cause.Error(),
-		HttpStatusCode: int32(httpStatusCode),
-		TransactionId:  int64(tx.ID), //nolint:gosec // overflow is acceptable for telemetry
-		AtlasUrl:       a.customURL.String(),
-		CreatedAt:      time.Now().UnixMicro(),
-	}
-
-	a.lggr.Infow("Emitting Atlas error event", "msg", msg)
-
-	messageBytes, err := proto.Marshal(msg)
-	if err != nil {
-		a.lggr.Errorw("Failed to marshal Atlas error event", "err", err)
-		return
-	}
-
-	attrKVs := []any{
-		"beholder_domain", "svr",
-		"beholder_entity", "svr.v1.FastLaneAtlasError",
-		"beholder_data_schema", "/fastlane-atlas-error/versions/1",
-	}
-
-	mStr := protojson.MarshalOptions{
-		UseProtoNames:   true,
-		EmitUnpopulated: true,
-	}.Format(msg)
-	a.lggr.Infow("[Beholder.emit]", "message", mStr, "attributes", attrKVs)
-
-	if emitErr := a.beholderEmitter.Emit(ctx, messageBytes, attrKVs...); emitErr != nil {
-		a.lggr.Errorw("Failed to emit Atlas error event", "err", emitErr)
-	}
-}
-
-// emitAtlasError emits an OTel event to track FastLane Atlas errors
-func (a *MetaClient) emitAtlasError(ctx context.Context, errType string, err error, tx *types.Transaction) {
-	a.emitAtlasErrorWithHttpStatusCode(ctx, errType, err, -1, tx)
 }
 
 func (a *MetaClient) SendTransaction(ctx context.Context, tx *types.Transaction, attempt *types.Attempt) error {
@@ -244,13 +173,13 @@ func (a *MetaClient) SendTransaction(ctx context.Context, tx *types.Transaction,
 		meta, err := a.SendRequest(ctx, tx, attempt, *meta.DualBroadcastParams, tx.ToAddress)
 		if err != nil {
 			a.metrics.RecordSendRequestError(ctx)
-			a.emitAtlasError(ctx, "send_request", err, tx)
+			a.metrics.emitAtlasError(ctx, "send_request", a.customURL, err, tx)
 			return fmt.Errorf("error sending request for transactionID(%d): %w", tx.ID, err)
 		}
 		if meta != nil {
 			if err := a.SendOperation(ctx, tx, attempt, *meta); err != nil {
 				a.metrics.RecordSendOperationError(ctx)
-				a.emitAtlasError(ctx, "send_operation", err, tx)
+				a.metrics.emitAtlasError(ctx, "send_operation", a.customURL, err, tx)
 				return fmt.Errorf("failed to send operation for transactionID(%d): %w", tx.ID, err)
 			}
 			return nil
@@ -388,21 +317,21 @@ func (a *MetaClient) SendRequest(parentCtx context.Context, tx *types.Transactio
 	signature, err := a.ks.SignMessage(parentCtx, tx.FromAddress, []byte(payload))
 	if err != nil {
 		wrappedErr := fmt.Errorf("failed to sign message: %w", err)
-		a.emitAtlasError(ctx, "sign_message", wrappedErr, tx)
+		a.metrics.emitAtlasError(ctx, "sign_message", a.customURL, wrappedErr, tx)
 		return nil, wrappedErr
 	}
 	params.Signature = signature
 	marshalledParamsExtended, err := json.Marshal(params)
 	if err != nil {
 		wrappedErr := fmt.Errorf("failed to marshal signed params: %w", err)
-		a.emitAtlasError(ctx, "marshal_params", wrappedErr, tx)
+		a.metrics.emitAtlasError(ctx, "marshal_params", a.customURL, wrappedErr, tx)
 		return nil, wrappedErr
 	}
 	body := fmt.Appendf(nil, `{"jsonrpc":"2.0","method":"%s","params":[%s], "id":1}`, string(m), marshalledParamsExtended)
 	req, err := http.NewRequestWithContext(ctx, http.MethodPost, a.customURL.String(), bytes.NewBuffer(body))
 	if err != nil {
 		wrappedErr := fmt.Errorf("failed to create POST request: %w", err)
-		a.emitAtlasError(ctx, "create_request", wrappedErr, tx)
+		a.metrics.emitAtlasError(ctx, "create_request", a.customURL, wrappedErr, tx)
 		return nil, wrappedErr
 	}
 	req.Header.Add("Content-Type", "application/json")
@@ -416,7 +345,7 @@ func (a *MetaClient) SendRequest(parentCtx context.Context, tx *types.Transactio
 	a.metrics.RecordLatency(ctx, latency)
 	if err != nil {
 		wrappedErr := fmt.Errorf("failed to send POST request: %w", err)
-		a.emitAtlasError(ctx, "http_request", wrappedErr, tx)
+		a.metrics.emitAtlasError(ctx, "http_request", a.customURL, wrappedErr, tx)
 		return nil, wrappedErr
 	}
 	defer resp.Body.Close()
@@ -426,20 +355,20 @@ func (a *MetaClient) SendRequest(parentCtx context.Context, tx *types.Transactio
 
 	if resp.StatusCode != http.StatusOK {
 		httpErr := fmt.Errorf("request %v failed with status: %d", req, resp.StatusCode)
-		a.emitAtlasErrorWithHttpStatusCode(ctx, "http_status", httpErr, resp.StatusCode, tx)
+		a.metrics.emitAtlasErrorWithHttpStatusCode(ctx, "http_status", a.customURL, httpErr, resp.StatusCode, tx)
 		return nil, httpErr
 	}
 
 	data, err := io.ReadAll(resp.Body)
 	if err != nil {
-		a.emitAtlasErrorWithHttpStatusCode(ctx, "read_response", err, resp.StatusCode, tx)
+		a.metrics.emitAtlasErrorWithHttpStatusCode(ctx, "read_response", a.customURL, err, resp.StatusCode, tx)
 		return nil, err
 	}
 
 	var response Response
 	err = json.Unmarshal(data, &response)
 	if err != nil {
-		a.emitAtlasErrorWithHttpStatusCode(ctx, "unmarshal_response", err, resp.StatusCode, tx)
+		a.metrics.emitAtlasErrorWithHttpStatusCode(ctx, "unmarshal_response", a.customURL, err, resp.StatusCode, tx)
 		return nil, err
 	}
 
@@ -449,7 +378,7 @@ func (a *MetaClient) SendRequest(parentCtx context.Context, tx *types.Transactio
 			return nil, nil
 		}
 		atlasErr := errors.New(response.Error.ErrorMessage)
-		a.emitAtlasErrorWithHttpStatusCode(ctx, "atlas_error", atlasErr, resp.StatusCode, tx)
+		a.metrics.emitAtlasErrorWithHttpStatusCode(ctx, "atlas_error", a.customURL, atlasErr, resp.StatusCode, tx)
 		return nil, atlasErr
 	}
 
@@ -588,7 +517,7 @@ func VerifyMetadata(txData []byte, fromAddress common.Address, result Metacallda
 func (a *MetaClient) SendOperation(ctx context.Context, tx *types.Transaction, attempt *types.Attempt, meta MetacalldataResponse) error {
 	if tx.Nonce == nil {
 		nonceErr := fmt.Errorf("failed to create attempt for txID: %v: nonce empty", tx.ID)
-		a.emitAtlasError(ctx, "nonce_empty", nonceErr, tx)
+		a.metrics.emitAtlasError(ctx, "nonce_empty", a.customURL, nonceErr, tx)
 		return nonceErr
 	}
 
@@ -600,7 +529,7 @@ func (a *MetaClient) SendOperation(ctx context.Context, tx *types.Transaction, a
 	gas := meta.GasLimit.ToInt()
 	if !gas.IsUint64() {
 		gasErr := fmt.Errorf("gas value does not fit in uint64: %s", gas)
-		a.emitAtlasError(ctx, "gas_overflow", gasErr, tx)
+		a.metrics.emitAtlasError(ctx, "gas_overflow", a.customURL, gasErr, tx)
 		return gasErr
 	}
 	dynamicTx := evmtypes.DynamicFeeTx{
@@ -615,13 +544,13 @@ func (a *MetaClient) SendOperation(ctx context.Context, tx *types.Transaction, a
 	signedTx, err := a.ks.SignTx(ctx, tx.FromAddress, evmtypes.NewTx(&dynamicTx))
 	if err != nil {
 		signErr := fmt.Errorf("failed to sign attempt for txID: %v, err: %w", tx.ID, err)
-		a.emitAtlasError(ctx, "sign_tx", signErr, tx)
+		a.metrics.emitAtlasError(ctx, "sign_tx", a.customURL, signErr, tx)
 		return signErr
 	}
 	a.lggr.Infow("Intercepted attempt for tx", "txID", tx.ID, "hash", signedTx.Hash(), "toAddress", meta.ToAddress, "gasLimit", meta.GasLimit,
 		"TipCap", tip, "FeeCap", meta.MaxFeePerGas)
 	if err := a.c.SendTransaction(ctx, signedTx); err != nil {
-		a.emitAtlasError(ctx, "send_tx", err, tx)
+		a.metrics.emitAtlasError(ctx, "send_tx", a.customURL, err, tx)
 		return err
 	}
 	return nil
