+++ pkg/txm/clientwrappers/dualbroadcast/meta_metrics_test.go
@@ -1,25 +1,39 @@
 package dualbroadcast
 
 import (
+	"context"
+	"errors"
+	"math/big"
+	"net/url"
 	"testing"
 	"time"
 
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/proto"
+
+	"github.com/smartcontractkit/chainlink-common/pkg/logger"
+	"github.com/smartcontractkit/chainlink-common/pkg/sqlutil"
+	"github.com/smartcontractkit/chainlink-evm/pkg/txm/types"
+	pb "github.com/smartcontractkit/chainlink-protos/svr/v1"
 )
 
 func TestMetaMetrics(t *testing.T) {
+	t.Parallel()
 	chainID := "1"
 
 	t.Run("NewMetaMetrics", func(t *testing.T) {
-		metrics, err := NewMetaMetrics(chainID)
+		metrics, err := NewMetaMetrics(chainID, logger.Test(t))
 		require.NoError(t, err)
 		assert.NotNil(t, metrics)
 		assert.Equal(t, chainID, metrics.chainID)
 	})
 
 	t.Run("RecordBasicMetrics", func(t *testing.T) {
-		metrics, err := NewMetaMetrics(chainID)
+		t.Parallel()
+		metrics, err := NewMetaMetrics(chainID, logger.Test(t))
 		require.NoError(t, err)
 
 		ctx := t.Context()
@@ -30,5 +44,184 @@ func TestMetaMetrics(t *testing.T) {
 		metrics.RecordBidsReceived(ctx, 5)
 		metrics.RecordSendRequestError(ctx)
 		metrics.RecordSendOperationError(ctx)
+	})
+}
+
+// mockBeholderEmitter is a mock for beholder.Emitter
+type mockBeholderEmitter struct {
+	mock.Mock
+}
+
+func (m *mockBeholderEmitter) Emit(ctx context.Context, body []byte, attrKVs ...any) error {
+	args := m.Called(ctx, body, attrKVs)
+	return args.Error(0)
+}
+
+func (m *mockBeholderEmitter) Close() error {
+	args := m.Called()
+	return args.Error(0)
+}
+
+func TestMetaClient_emitAtlasErrorWithHttpStatusCode(t *testing.T) {
+	t.Parallel()
+	testChainID := big.NewInt(1)
+	testURL, _ := url.Parse("https://atlas.example.com")
+	lggr := logger.Test(t)
+
+	t.Run("emits error with all fields populated", func(t *testing.T) {
+		t.Parallel()
+		mockEmitter := new(mockBeholderEmitter)
+		metrics, err := NewMetaMetrics(testChainID.String(), lggr)
+		require.NoError(t, err)
+		metrics.emitter = mockEmitter
+
+		u, err := url.Parse("https://example.com")
+		require.NoError(t, err)
+
+		nonce := uint64(450)
+
+		tx := &types.Transaction{
+			ID:          123,
+			Nonce:       &nonce,
+			FromAddress: common.HexToAddress("0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+			ToAddress:   common.HexToAddress("0xBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB"),
+			// Meta:        nil,
+		}
+
+		var capturedBody []byte
+		mockEmitter.On("Emit", mock.Anything, mock.Anything, mock.Anything).
+			Run(func(args mock.Arguments) {
+				capturedBody = args.Get(1).([]byte)
+			}).
+			Return(nil)
+
+		metrics.emitAtlasErrorWithHttpStatusCode(t.Context(), "send_request", u, errors.New("test error message"), 500, tx)
+
+		mockEmitter.AssertExpectations(t)
+
+		// Verify the emitted message
+		var emittedMsg pb.FastLaneAtlasError
+		err = proto.Unmarshal(capturedBody, &emittedMsg)
+		require.NoError(t, err)
+
+		assert.Equal(t, testChainID.String(), emittedMsg.ChainId)
+		assert.Equal(t, tx.FromAddress.Hex(), emittedMsg.FromAddress)
+		assert.Equal(t, tx.ToAddress.Hex(), emittedMsg.ToAddress)
+		// assert.Equal(t, fwdrDestAddress.String(), emittedMsg.FeedAddress)
+		assert.Equal(t, "42", emittedMsg.Nonce)
+		assert.Equal(t, "test_error_type", emittedMsg.ErrorType)
+		assert.Equal(t, "test error message", emittedMsg.ErrorMessage)
+		assert.Equal(t, int32(500), emittedMsg.HttpStatusCode)
+		assert.Equal(t, int64(123), emittedMsg.TransactionId)
+		assert.Equal(t, testURL.String(), emittedMsg.AtlasUrl)
+	})
+
+	t.Run("emits error with nil nonce", func(t *testing.T) {
+		t.Parallel()
+		mockEmitter := new(mockBeholderEmitter)
+		metrics, err := NewMetaMetrics(testChainID.String(), lggr)
+		require.NoError(t, err)
+		metrics.emitter = mockEmitter
+
+		tx := &types.Transaction{
+			ID:          456,
+			Nonce:       nil, // nil nonce
+			FromAddress: common.HexToAddress("0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+			ToAddress:   common.HexToAddress("0xBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB"),
+			Meta:        nil,
+		}
+
+		var capturedBody []byte
+		mockEmitter.On("Emit", mock.Anything, mock.Anything, mock.Anything).
+			Run(func(args mock.Arguments) {
+				capturedBody = args.Get(1).([]byte)
+			}).
+			Return(nil)
+
+		metrics.emitAtlasErrorWithHttpStatusCode(t.Context(), "error_type", testURL, errors.New("some error"), 400, tx)
+
+		mockEmitter.AssertExpectations(t)
+
+		var emittedMsg pb.FastLaneAtlasError
+		err = proto.Unmarshal(capturedBody, &emittedMsg)
+		require.NoError(t, err)
+
+		assert.Equal(t, "", emittedMsg.Nonce)       // empty string when nonce is nil
+		assert.Equal(t, "", emittedMsg.FeedAddress) // empty string when meta is nil
+	})
+
+	t.Run("emits error with negative http status code for non-http errors", func(t *testing.T) {
+		t.Parallel()
+		mockEmitter := new(mockBeholderEmitter)
+		metrics, err := NewMetaMetrics(testChainID.String(), lggr)
+		require.NoError(t, err)
+		metrics.emitter = mockEmitter
+
+		tx := &types.Transaction{
+			ID:          789,
+			FromAddress: common.HexToAddress("0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+			ToAddress:   common.HexToAddress("0xBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB"),
+		}
+
+		var capturedBody []byte
+		mockEmitter.On("Emit", mock.Anything, mock.Anything, mock.Anything).
+			Run(func(args mock.Arguments) {
+				capturedBody = args.Get(1).([]byte)
+			}).
+			Return(nil)
+
+		metrics.emitAtlasErrorWithHttpStatusCode(t.Context(), "non_http_error", testURL, errors.New("network error"), -1, tx)
+
+		mockEmitter.AssertExpectations(t)
+
+		var emittedMsg pb.FastLaneAtlasError
+		err = proto.Unmarshal(capturedBody, &emittedMsg)
+		require.NoError(t, err)
+
+		assert.Equal(t, int32(-1), emittedMsg.HttpStatusCode)
+	})
+
+	t.Run("handles emit error gracefully", func(t *testing.T) {
+		t.Parallel()
+		mockEmitter := new(mockBeholderEmitter)
+		metrics, err := NewMetaMetrics(testChainID.String(), lggr)
+		require.NoError(t, err)
+		metrics.emitter = mockEmitter
+
+		tx := &types.Transaction{
+			ID:          999,
+			FromAddress: common.HexToAddress("0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+			ToAddress:   common.HexToAddress("0xBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB"),
+		}
+
+		mockEmitter.On("Emit", mock.Anything, mock.Anything, mock.Anything).
+			Return(errors.New("emit failed"))
+
+		// Should not panic, just log the error
+		metrics.emitAtlasErrorWithHttpStatusCode(t.Context(), "error_type", testURL, errors.New("some error"), 500, tx)
+
+		mockEmitter.AssertExpectations(t)
+	})
+
+	t.Run("handles invalid meta JSON gracefully", func(t *testing.T) {
+		t.Parallel()
+		mockEmitter := new(mockBeholderEmitter)
+		metrics, err := NewMetaMetrics(testChainID.String(), lggr)
+		require.NoError(t, err)
+		metrics.emitter = mockEmitter
+
+		invalidJSON := sqlutil.JSON([]byte("invalid json"))
+		tx := &types.Transaction{
+			ID:          111,
+			FromAddress: common.HexToAddress("0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+			ToAddress:   common.HexToAddress("0xBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB"),
+			Meta:        &invalidJSON,
+		}
+
+		// Should not call Emit because GetMeta will fail
+		metrics.emitAtlasErrorWithHttpStatusCode(t.Context(), "error_type", testURL, errors.New("some error"), 500, tx)
+
+		// Emit should not be called when meta parsing fails
+		mockEmitter.AssertNotCalled(t, "Emit", mock.Anything, mock.Anything, mock.Anything)
 	})
 }
